Replace atoi -> std::stoi
git-svn-id: https://subversion.gsi.de/dabc/trunk@4722 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ base/run/dabc_exe.cxx
@@ -14,6 +14,7 @@
  ************************************************************/
 
 #include <cstdlib>
+#include <string>
 #include <clocale>
 
 #include "dabc/Manager.h"
@@ -106,11 +107,11 @@ int main(int numc, char* args[])
       } else
       if (strcmp(arg,"-nodeid")==0) {
          if (cnt < numc)
-            nodeid = (unsigned) atoi(args[cnt++]);
+            nodeid = std::stoul(args[cnt++]);
       } else
       if (strcmp(arg,"-numnodes")==0) {
          if (cnt < numc)
-            numnodes = (unsigned) atoi(args[cnt++]);
+            numnodes = std::stoul(args[cnt++]);
       } else
       if (strcmp(arg,"-run")==0) {
          //dorun = true;
